:white_check_mark: migrate utilities based tests from `tests.py` to `tests_utilities.py`
